Changing boss health acquisition method
This changes the way the boss health is acquired for the field, mini
and main bosses. The boss names and health are in their respective csv
files and the program now uses a pandas dataframe to get a sample from
the provided data to acquire the boss name and health.

The *-boss-list.txt files have been renamed with the .csv extension.

A new except block has been added to catch IndexError if the correct
amount of data is not given to the program from the file.