Adding functionality for players to earn runes
Adds the functionality for the players in the game to earn runes for
defeating bosses. Runes are used as the currency that will allow players
to purchase stat upgrades (WIP). Each player (host and summons) earn
the same amount of runes from each boss encounter and the number of
runes a boss drops is determined when its stats are loaded from the csv
file for its respective category.

Adds the get_runes method for the Boss class. Adds the get_runes,
add_runes, and reduce_runes method for the Character class.

Also comes with various edits for easier readability of the output.