Update weapon data getter
Changes the way the init and update_stats functions
check for the weapon types before increasing the
player's attack or armor. Adds a list of weapon
types and a list of shield types to use for
checking the the item in the player's left hand
to determine if it is a weapon or a shield.

Removes depracated entries in the weapons files.